+++ app/src/main/java/com/unionbankph/corporate/loan/calculator/LoanInfoItem.kt
@@ -1,10 +1,10 @@
 package com.unionbankph.corporate.loan.calculator
 
 data class LoanInfoItem(
-    val loanAmount: Int? = null,
+    val loanAmount: Float? = null,
     val loanTenure: Int? = null,
-    var annualInterestRate: Int? = null,
-    var monthlyPayment: Int? = null,
-    var totalInterestPayable: Int? = null,
-    var totalAmountPayable: Int? = null
+    var annualInterestRate: Float? = null,
+    var monthlyPayment: Float? = null,
+    var totalInterestPayable: Float? = null,
+    var totalAmountPayable: Float? = null
 )
